[CodeGen] using abi encode for function returns
