+++ include/soll/AST/TypePtr.h
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+#pragma once
 #include "soll/AST/Type.h"
 
 namespace soll {
